+++ src/index.ts
@@ -7,7 +7,7 @@ import ConfigureProgram from "./programs/configure/configure-program.js";
 import TranslateProgram from "./programs/translate-program.js";
 import UnderstandProgram from "./programs/understand-program.js";
 
-const version = "0.1.1";
+const version = "0.1.2";
 const description =
   "A super charged CLI for interfacing with GPT-3 and other AI services";